Rename exposureconfig package
