+++ services/distribution/src/main/java/app/coronawarn/server/services/distribution/assembly/appconfig/ExposureConfigurationProvider.java
@@ -17,18 +17,9 @@
  * under the License.
  */
 
-package app.coronawarn.server.services.distribution.assembly.exposureconfig;
+package app.coronawarn.server.services.distribution.assembly.appconfig;
 
 import app.coronawarn.server.common.protocols.internal.RiskScoreParameters;
-import app.coronawarn.server.common.protocols.internal.RiskScoreParameters.Builder;
-import app.coronawarn.server.services.distribution.assembly.exposureconfig.parsing.YamlConstructorForProtoBuf;
-import java.io.IOException;
-import java.io.InputStream;
-import org.springframework.core.io.ClassPathResource;
-import org.springframework.core.io.Resource;
-import org.yaml.snakeyaml.Yaml;
-import org.yaml.snakeyaml.error.YAMLException;
-import org.yaml.snakeyaml.introspector.BeanAccess;
 
 /**
  * Provides the Exposure Configuration based on a file in the file system.<br> The existing file must be a valid YAML
@@ -62,21 +53,6 @@ public class ExposureConfigurationProvider {
    * @throws UnableToLoadFileException when the file/transformation did not succeed
    */
   public static RiskScoreParameters readFile(String path) throws UnableToLoadFileException {
-    Yaml yaml = new Yaml(new YamlConstructorForProtoBuf());
-    yaml.setBeanAccess(BeanAccess.FIELD); /* no setters on RiskScoreParameters available */
-
-    Resource riskScoreParametersResource = new ClassPathResource(path);
-    try (InputStream inputStream = riskScoreParametersResource.getInputStream()) {
-      Builder loaded = yaml.loadAs(inputStream, RiskScoreParameters.newBuilder().getClass());
-      if (loaded == null) {
-        throw new UnableToLoadFileException(path);
-      }
-
-      return loaded.build();
-    } catch (YAMLException e) {
-      throw new UnableToLoadFileException("Parsing failed", e);
-    } catch (IOException e) {
-      throw new UnableToLoadFileException("Failed to load file " + path, e);
-    }
+    return YamlLoader.loadYamlIntoProtobufBuilder(path, RiskScoreParameters.Builder.class).build();
   }
 }
