+++ services/distribution/src/main/java/app/coronawarn/server/services/distribution/assembly/appconfig/validation/ParameterSpec.java
@@ -17,21 +17,30 @@
  * under the License.
  */
 
-package app.coronawarn.server.services.distribution.assembly.exposureconfig.validation;
+package app.coronawarn.server.services.distribution.assembly.appconfig.validation;
 
 /**
- * Classes that extend {@link AppConfigurationValidator} validate the values of an associated {@link
- * com.google.protobuf.Message} instance.
+ * Definition of the spec according to Apple/Google:
+ * https://developer.apple.com/documentation/exposurenotification/enexposureconfiguration
  */
-public abstract class AppConfigurationValidator {
+public class ParameterSpec {
 
-  protected ValidationResult errors;
+  private ParameterSpec() {
+  }
 
   /**
-   * Performs a validation of the associated {@link com.google.protobuf.Message} instance and returns information about
-   * validation failures.
-   *
-   * @return The ValidationResult instance, containing information about possible errors.
+   * The minimum weight value for mobile API.
    */
-  public abstract ValidationResult validate();
+  public static final double WEIGHT_MIN = 0.001;
+
+  /**
+   * The maximum weight value for mobile API.
+   */
+  public static final int WEIGHT_MAX = 100;
+
+  /**
+   * Maximum number of allowed decimals.
+   */
+  public static final int WEIGHT_MAX_DECIMALS = 3;
+
 }
