+++ services/distribution/src/main/java/app/coronawarn/server/services/distribution/assembly/appconfig/validation/ValidationError.java
@@ -17,30 +17,11 @@
  * under the License.
  */
 
-package app.coronawarn.server.services.distribution.assembly.exposureconfig.validation;
+package app.coronawarn.server.services.distribution.assembly.appconfig.validation;
 
 /**
- * Definition of the spec according to Apple/Google:
- * https://developer.apple.com/documentation/exposurenotification/enexposureconfiguration
+ * A validation error, found during validation of a generated protocol buffers message class instance.
  */
-public class ParameterSpec {
-
-  private ParameterSpec() {
-  }
-
-  /**
-   * The minimum weight value for mobile API.
-   */
-  public static final double WEIGHT_MIN = 0.001;
-
-  /**
-   * The maximum weight value for mobile API.
-   */
-  public static final int WEIGHT_MAX = 100;
-
-  /**
-   * Maximum number of allowed decimals.
-   */
-  public static final int WEIGHT_MAX_DECIMALS = 3;
+public interface ValidationError {
 
 }
